Changed the strategy to connect the sales and purchase workflows:
Now when the sales order is confirmed, we create the procurement (with a link to the PO),
but we do not anymore confirm it. We leave it as 'draft'. Only once the picking of the purchase
order is created, we link the move of the picking with the procurement and confirm it.

This is necessary, because otherwise the procurement creates a new move
(procurement -> customer) when we confirm it.