fix: don't set code on PO generated from agreement
When generating a purchase order from a framework agreement, the code
should be auto-generated from the sequence, and not be equal to the code
of the logistic requisition.

This bug was already fixed on v7 by Nicolas Bessi. Then for some reason,
while porting to v8 I re-enabled a commented out line bringing that
back.

Now it is deleted for good.