compute total currency only when data is available
Fixes #33.

This is due to the fact that with the v8 API the methods that compute
function fields are called not only on save, but also as automatic
onchanges. In that situation, `self` is not a true record, but a virtual
NewId that contains only the values present in the view.

That is cool in theory, but in cases like that it can crash because the
requisition_id is empty as it is not present in the view.

This fix does the computation that relies on requisition_id only when it
is available. I do not check explicitly whether self if a NewId, so that
if we manage to fix the view so that we know the requisition_id, the
amount_total_company will be updated on the fly.