remove existing tests: they are now irrelevant
I discussed this with @yvaucher and @rdeheele. These tests are long
flows taken from logistic_requisition before logistic_budget was split
out, that check functionality of logistic_budget that is not there
anymore (we removed the old constraints). We did not notice the problem
because the only assertion was disabled to work around a bug in the
core, recently fixed.

For the new functionality I will not recover these tests, as they are
not really relevant here, and add new ones, more isolated.