fix: restore onchange on tender line form
This gives to the product field the same onchange we have in the tree
view.

One side effect of this is that by default this sets the quantity to 1
(instead of zero). Lines with quantity zero are easily overseen and give
problems later when bids cannot be selected because of differing
quantities.